requirements.txt
alembic==1.7.6
aniso8601==9.0.1
Babel==2.12.1
blinker==1.6.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
elastic-transport==8.4.0
elasticsearch==8.8.2
Flask==2.0.2
Flask-Admin==1.6.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.5
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
jinja-to-js==3.2.3
Jinja2==3.1.2
Mako==1.1.6
MarkupSafe==2.0.1
packaging==23.1
Pillow==9.1.1
psycopg2-binary
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
